requirements.txt
dlib>=19.24
face-recognition>=1.3.0
numpy>=1.26.0
Pillow>=10.0.0
scipy>=1.11.0
fastapi>=0.111.0
uvicorn>=0.30.1
python-multipart>=0.0.9
scikit-learn>=1.3.0
insightface>=0.7.3
onnxruntime>=1.15.1
